Add a preliminary set of API functions
It looks like there are a few internal quirks (e.g. loosely coupled
class attributes for elements, their names and their namespaces)
that make incremental creation of records kind of cumbersome.
I need to investigate if I'm using PRANG the wrong way, or if PRANG
needs to be improved, or if I should go for a different module
(XML::Compile::Schema perhaps or XML::Toolkit).